pyproject.toml
[project]
name = "groundstation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiohttp>=3.13.3",
    "ephem>=4.2",
    "jinja2>=3.1.6",
    "pydantic>=2.9",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "zstandard>=0.25.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
]
